feat: VoteService에 최신 공개 투표에 참여 메서드 추가 (#368)
